requirements.txt
google-cloud-firestore==2.20.0
fastapi==0.115.8
uvicorn==0.34.0
PyJWT==2.10.1
python-dotenv==1.0.1
passlib==1.7.4
jose==1.0.0
pycryptodome==3.21.0
